+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -273,6 +273,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Change how method for specular occlusions is decided on inspector shader (Lit, LitTesselation, LayeredLit, LayeredLitTessellation)
 - Unlocked SSS, SSR, Motion Vectors and Distortion frame settings for reflections probes.
 - Changed the ownership of temporary ray tracing buffers to reduce the memory footprint.
+- Removed dead code related to stencil handling. 
 
 ## [7.1.1] - 2019-09-05